[Fixes #168] [SOS] add endpoint to retrieve SOS and observable properties
